+++ src/Queue.js
@@ -296,23 +296,23 @@ Queue.prototype.peek = function peek(ignoreDelivered) {
   }
 };
 
-Queue.prototype.cancel = function cancel(consumerTag) {
+Queue.prototype.cancel = function cancel(consumerTag, requeue) {
   const consumers = this[kConsumers];
   const idx = consumers.findIndex((c) => c.consumerTag === consumerTag);
   if (idx === -1) return;
 
   const consumer = consumers[idx];
-  this.unbindConsumer(consumer);
+  this.unbindConsumer(consumer, requeue);
 };
 
-Queue.prototype.dismiss = function dismiss(onMessage) {
+Queue.prototype.dismiss = function dismiss(onMessage, requeue) {
   const consumers = this[kConsumers];
   const consumer = consumers.find((c) => c.onMessage === onMessage);
   if (!consumer) return;
-  this.unbindConsumer(consumer);
+  this.unbindConsumer(consumer, requeue);
 };
 
-Queue.prototype.unbindConsumer = function unbindConsumer(consumer) {
+Queue.prototype.unbindConsumer = function unbindConsumer(consumer, requeue = true) {
   const consumers = this[kConsumers];
   const idx = consumers.indexOf(consumer);
   if (idx === -1) return;
@@ -322,7 +322,7 @@ Queue.prototype.unbindConsumer = function unbindConsumer(consumer) {
   this[kExclusive] = false;
 
   consumer.stop();
-  consumer.nackAll(true);
+  consumer.nackAll(requeue);
 
   this.emit('consumer.cancel', consumer);